+++ samples/xskrestricted/bpf/xskrestricted_redirect.c
@@ -21,7 +21,7 @@
 struct
 {
     __uint(type, BPF_MAP_TYPE_XSKMAP);
-    __type(key, uint32_t);
+    __type(key, uint64_t);
     __type(value, void *);
     __uint(max_entries, 64);
 } xsk_map SEC(".maps");
