+++ test/bpf/xsk_redirect.c
@@ -14,7 +14,7 @@
 struct
 {
     __uint(type, BPF_MAP_TYPE_XSKMAP);
-    __type(key, uint32_t);
+    __type(key, uint64_t);
     __type(value, void *);
     __uint(max_entries, 64);
 } xsk_map SEC(".maps");
@@ -23,6 +23,6 @@ SEC("xdp/xsk_redirect")
 int
 xsk_redirect(xdp_md_t *ctx)
 {
-    uint32_t index = ctx->rx_queue_index;
+    uint64_t index = ctx->rx_queue_index;
     return bpf_redirect_map(&xsk_map, index, XDP_PASS);
 }
