+++ test/bpf/xsk_redirect_fallback.c
@@ -10,7 +10,7 @@
 struct
 {
     __uint(type, BPF_MAP_TYPE_XSKMAP);
-    __type(key, uint32_t);
+    __type(key, uint64_t);
     __type(value, void *);
     __uint(max_entries, 64);
 } xsk_map SEC(".maps");
@@ -32,7 +32,7 @@ SEC("xdp/xsk_redirect_fallback")
 int
 xsk_redirect_fallback(xdp_md_t *ctx)
 {
-    uint32_t index = ctx->rx_queue_index;
+    uint64_t index = ctx->rx_queue_index;
     uint32_t zero = 0;
     uint64_t fallback = XDP_PASS;
